+++ ignite/cmd/scaffold.go
@@ -228,7 +228,7 @@ func scaffoldType(
 		return err
 	}
 
-	if err := sc.PostScaffold(cmd.Context(), cacheStorage); err != nil {
+	if err := sc.PostScaffold(cmd.Context(), cacheStorage, false); err != nil {
 		return err
 	}