+++ ignite/cmd/scaffold_chain.go
@@ -85,6 +85,7 @@ about Cosmos SDK on https://docs.cosmos.network
 	c.Flags().StringSlice(flagParams, []string{}, "add default module parameters")
 	c.Flags().StringSlice(flagModuleConfigs, []string{}, "add module configs")
 	c.Flags().Bool(flagSkipGit, false, "skip Git repository initialization")
+	c.Flags().Bool(flagSkipProto, false, "skip proto generation")
 	c.Flags().Bool(flagMinimal, false, "create a minimal blockchain (with the minimum required Cosmos SDK modules)")
 	c.Flags().Bool(flagIsConsumer, false, "scafffold an ICS consumer chain")
 	// Cannot have both minimal and consumer flag
@@ -107,6 +108,7 @@ func scaffoldChainHandler(cmd *cobra.Command, args []string) error {
 		isConsumer, _      = cmd.Flags().GetBool(flagIsConsumer)
 		params, _          = cmd.Flags().GetStringSlice(flagParams)
 		moduleConfigs, _   = cmd.Flags().GetStringSlice(flagModuleConfigs)
+		skipProto, _       = cmd.Flags().GetBool(flagSkipProto)
 	)
 
 	if noDefaultModule {
@@ -148,7 +150,7 @@ func scaffoldChainHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	if err := scaffolder.PostScaffold(cmd.Context(), cacheStorage, appDir, goModule); err != nil {
+	if err := scaffolder.PostScaffold(cmd.Context(), cacheStorage, appDir, goModule, skipProto); err != nil {
 		return err
 	}