+++ ignite/cmd/scaffold_configs.go
@@ -73,7 +73,7 @@ func scaffoldConfigsHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	if err := sc.PostScaffold(cmd.Context(), cacheStorage); err != nil {
+	if err := sc.PostScaffold(cmd.Context(), cacheStorage, false); err != nil {
 		return err
 	}