+++ ignite/cmd/scaffold_message.go
@@ -130,7 +130,7 @@ func messageHandler(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	if err := sc.PostScaffold(cmd.Context(), cacheStorage); err != nil {
+	if err := sc.PostScaffold(cmd.Context(), cacheStorage, false); err != nil {
 		return err
 	}