+++ ignite/services/scaffolder/scaffolder.go
@@ -63,14 +63,15 @@ func New(appPath string) (Scaffolder, error) {
 	return s, nil
 }
 
-func (s Scaffolder) PostScaffold(ctx context.Context, cacheStorage cache.Storage) error {
-	return PostScaffold(ctx, cacheStorage, s.path, s.modpath.RawPath)
+func (s Scaffolder) PostScaffold(ctx context.Context, cacheStorage cache.Storage, skipProto bool) error {
+	return PostScaffold(ctx, cacheStorage, s.path, s.modpath.RawPath, skipProto)
 }
 
-func PostScaffold(ctx context.Context, cacheStorage cache.Storage, path, gomodPath string) error {
-	err := protoc(ctx, cacheStorage, path, gomodPath)
-	if err != nil {
-		return err
+func PostScaffold(ctx context.Context, cacheStorage cache.Storage, path, gomodPath string, skipProto bool) error {
+	if !skipProto {
+		if err := protoc(ctx, cacheStorage, path, gomodPath); err != nil {
+			return err
+		}
 	}
 
 	if err := gocmd.ModTidy(ctx, path); err != nil {